integrate the two CLIRunner classes
